Remove need for commons-io
Files.copy seems to be a viable alternative for IOUtils.copy
at least when it comes to writing stream to file.

Also, the commons-io dependency causes some issues for projects
using discord-same-sdk4j as a dependencies, as it makes it look
like commons-io is available while in reality it is not.