Fix issues on Windows caused by different integer widths
Apparently when compiling on Linux for Windows (with mingw)
"long long int" is not the same as "long int" (aka long).
This made the compiler produce some warnings and more
importantly caused some longs (which I expected to be 64 bit)
to be shortened to something just over 32 bit.
Due to that, some IDs (of Lobbies and users), which are passed
to the event handler methods, were incorrect.